# lib/api-zod/src/generated/types/orderList.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Subscriptions Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderSummary } from './orderSummary';

export interface OrderList {
  orders: OrderSummary[];
  total: number;
  page: number;
  limit: number;
}