# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Subscriptions Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** e.g. "ChatGPT Plus", "Claude Pro" */
  service: string;
  /** Price in Telegram Stars for 1 month */
  price1m: number;
  /**
     * Price for 3 months
     * @nullable
     */
  price3m?: number | null;
  /**
     * Price for 6 months
     * @nullable
     */
  price6m?: number | null;
  /**
     * Price for 12 months
     * @nullable
     */
  price12m?: number | null;
  active: boolean;
  createdAt: Date;
}