# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Subscriptions Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  service: string;
  price1m: number;
  /** @nullable */
  price3m?: number | null;
  /** @nullable */
  price6m?: number | null;
  /** @nullable */
  price12m?: number | null;
  active?: boolean;
}