# lib/api-zod/src/generated/types/revenuePoint.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Subscriptions Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface RevenuePoint {
  date: string;
  stars: number;
  orders: number;
}