# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Subscriptions Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  totalOrders: number;
  totalSpentStar: number;
  createdAt: Date;
}